Loading now works with the new interface, and even handles the memory leak which turned out to be much easier to fix than I thought it would.

+++ src/model/gamemap.cpp
@@ -589,8 +589,7 @@ bool GameMap::isRowColInMapBounds(const int& row, const int& col) const {
 
 bool GameMap::isIndexInMapBounds(const int& index) const {
 
-    // warning C4018: '>=': signed/unsigned mismatch
-    if (index >= tiles.size()) {
+    if (index >= static_cast<int>(tiles.size())) {
         return false;
     }
